Move constants file because index.ts made me mad
