+++ example/app/(tabs)/flashlist.tsx
@@ -1,5 +1,5 @@
 import renderItem from '@/app/renderItem';
-import { DO_SCROLL_TEST, DRAW_DISTANCE, ESTIMATED_ITEM_LENGTH, RECYCLE_ITEMS } from '@/constants';
+import { DO_SCROLL_TEST, DRAW_DISTANCE, ESTIMATED_ITEM_LENGTH, RECYCLE_ITEMS } from '@/constants/constants';
 import { useScrollTest } from '@/constants/useScrollTest';
 import { FlashList, ListRenderItemInfo } from '@shopify/flash-list';
 import { Fragment, useRef } from 'react';
