+++ example/constants/constants.ts
@@ -1,5 +1,5 @@
 export const RECYCLE_ITEMS = false;
-export const EMULATE_SLOW_RENDER = 2;
-export const DRAW_DISTANCE = 5000;
+export const EMULATE_SLOW_RENDER = 5;
+export const DRAW_DISTANCE = 1000;
 export const DO_SCROLL_TEST = false;
 export const ESTIMATED_ITEM_LENGTH = 400;
