+++ src/constants.ts
@@ -1,3 +1,3 @@
 export const DEFAULT_SCROLL_BUFFER = 0;
 export const POSITION_OUT_OF_VIEW = -10000;
-export const OPTIMIZE_DIRECTION = true;
+export const OPTIMIZE_DIRECTION = false;
